Gps Track drawing

+++ app/src/main/java/com/example/summitdiary/database/Hike.kt
@@ -7,7 +7,7 @@ data class Hike(
     @PrimaryKey(autoGenerate = true) val hike_id: Long = 0,
     var title: String,
     var date: String,
-    var distance: Int,
+    var distance: Double,
     var time: String,
     var place_id: Long,
     var gpx_path: String
